+++ NEWS.md
@@ -1,4 +1,7 @@
 # speedrunr 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
-* Added `get_game`. Search for a game and retrieve its `id` which is required for the next step.
+* Added functions:
+    - `get_games`: Search for a game and retrieve its `id` which is required for the next step.
+    - `get_platforms`: List all the platforms runs are being done on.
+    - `get_categories`: Get a game's categories (needs `id` as per `get_games`).